runtime/secrets: separate API reading from Secret conversion
Separate file responsibilities in the runtime/secrets package:
- reader.go: API server reading functions (XXXFromSecretRef)
- converter.go: Secret conversion functions (XXXFromSecret)

The client-based API variants (TLSConfigFromSecretRef,
ProxyURLFromSecretRef) provide convenient auto-fetching for
use cases where Secrets contain only relevant data without
mixed authentication information.
